practice-course-udemy/scrapers/requirements.txt
requests-html
urllib3
beautifulsoup4
